+++ src/widgets/Sidebar.tsx
@@ -1,15 +1,16 @@
 import styled from 'styled-components'
 import React from 'react'
+import { deleteVote } from '../services/FirestoreApi'
 
 const FSidebar = styled.div`
   grid-area: sidebar;
-  padding: 14px;
+  /* padding: 14px; */
   color: lightgrey;
   background: #051e34;
 `
 
 const FLogo = styled.div`
-  margin-bottom: 5px;
+  margin: 11px 14px 10px 14px;
   font-family: 'Roboto', sans-serif;
   font-weight: 400;
   color: white;
@@ -17,17 +18,24 @@ const FLogo = styled.div`
 `
 
 const FMenuHeader = styled.div`
-  margin-top: 10px;
-  padding: 10px 0;
-  font-size: 17px;
+  /* margin: 14px; */
+  padding: 12px 14px;
+  font-size: 16px;
   font-family: 'Roboto', sans-serif;
   font-weight: 500;
+  background-color: #122c44;
 `
 
 const FMenuItem = styled.div`
-  padding: 8px 0;
-  font-size: 15px;
+  padding: 8px 14px;
+  font-size: 14px;
   font-family: 'Roboto', sans-serif;
+  background-color: #122c44;
+
+  &:hover {
+    background-color: #253d53;
+    cursor: pointer;
+  }
 `
 
 type TSidebar = {
@@ -36,6 +44,12 @@ type TSidebar = {
 }
 
 const Sidebar = (props: TSidebar) => {
+  function handleResetVoting () {
+    const voteId = 'wkNnqCPmR3mFIaUffIrC' // TODO: Actually impl this! 4/22/21
+    deleteVote(voteId)
+    console.log('>> Vote deleted:', voteId)
+  }
+
   const content = props.collapseSidebar
     ? <>
         <FLogo>📗</FLogo>
@@ -51,7 +65,7 @@ const Sidebar = (props: TSidebar) => {
         <FMenuHeader>Admin Tools</FMenuHeader>
         <FMenuItem>⏲️ Restart Countdown</FMenuItem>
         <FMenuItem>📜 Generate Results</FMenuItem>
-        <FMenuItem>🗳️ Reset Voting</FMenuItem>
+        <FMenuItem onClick={handleResetVoting}>🗳️ Reset Voting</FMenuItem>
       </>
 
   function handleClick () {
@@ -61,7 +75,7 @@ const Sidebar = (props: TSidebar) => {
   return (
     <FSidebar>
       {content}
-      <br />
+      <hr />
       <button onClick={handleClick}>c</button>
     </FSidebar>
   )